move all tiny components into their own files
