+++ packages/client/src/game/BulletCtor.ts
@@ -1,0 +1,5 @@
+import Bullet from "./Bullet";
+
+export default interface BulletCtor {
+  new (x: number, y: number, angle: number): Bullet;
+}
