+++ packages/client/src/game/Display.ts
@@ -1,0 +1,5 @@
+import Game from "./Game";
+
+export default interface Display {
+  draw(game: Game): void;
+}
