+++ packages/client/src/game/Game.ts
@@ -1,358 +1,19 @@
 import type TileRegistration from "@smiley-face-game/api/tiles/TileRegistration";
-import type { ZSPacket, ZSInit, ZSEvent, ZSPlayerJoin, ZWorldAction } from "@smiley-face-game/api/packets";
-import { TileLayer, ZWorldActionKindReply } from "@smiley-face-game/api/types";
-import { bresenhamsLine } from "@smiley-face-game/api/misc";
-
-export interface Position {
-  x: number;
-  y: number;
-}
-
-export type Velocity = Position;
-
-export interface Size {
-  width: number;
-  height: number;
-}
-
-export interface Inputs {
-  jump: boolean;
-  up: boolean;
-  down: boolean;
-  left: boolean;
-  right: boolean;
-}
-
-export enum GunState {
-  None,
-  Carrying,
-  Held,
-}
-
-export interface PhysicsObject {
-  position: Position;
-  velocity: Velocity;
-  tick?: (game: Game) => void;
-}
-
-interface BulletCtor {
-  new (x: number, y: number, angle: number): Bullet;
-}
-export class Bullet implements PhysicsObject {
-  creation: Date = new Date();
-  position: Position;
-  velocity: Velocity;
-
-  constructor(x: number, y: number, angle: number) {
-    const startingPower = 32;
-    this.position = { x, y };
-    this.velocity = {
-      x: Math.cos(angle) * startingPower,
-      y: Math.sin(angle) * startingPower,
-    };
-  }
-
-  tick() {
-    this.position.x += this.velocity.x;
-    this.position.y += this.velocity.y;
-  }
-}
-
-export function defaultInputs(): Inputs {
-  return { jump: false, up: false, down: false, left: false, right: false };
-}
-
-interface PlayerCtor {
-  new (id: number, username: string, isGuest: boolean): Player;
-}
-
-export class Physics {}
-
-export class Player implements PhysicsObject {
-  position: Position = { x: 0, y: 0 };
-  velocity: Velocity = { x: 0, y: 0 };
-  input: Inputs = defaultInputs();
-  role: "non" | "edit" | "staff" | "owner" = "non"; // TODO: remove role in favor of permission based stuff
-  private gunState: GunState = 0;
-
-  get hasGun(): boolean {
-    return this.gunState >= GunState.Carrying;
-  }
-
-  get isGunHeld(): boolean {
-    return this.gunState === GunState.Held;
-  }
-
-  get hasEdit(): boolean {
-    return this.role === "edit" || this.role === "owner";
-  }
-
-  constructor(readonly id: number, readonly username: string, readonly isGuest: boolean) {}
-
-  pickupGun() {
-    if (this.hasGun) throw new Error("picked up gun when already have a gun");
-    this.gunState = GunState.Carrying;
-  }
-
-  holdGun(isHeld: boolean) {
-    if (!this.hasGun) throw new Error("can't hold gun when there is no gun");
-    this.gunState = isHeld ? GunState.Held : GunState.Carrying;
-  }
-
-  tick(game: Game) {
-    // stupidly simple physics just so we can prototype for now
-    // fancy stuff coming later <3
-    this.velocity.x = 0;
-    this.velocity.y = 0;
-    if (this.input.up) this.velocity.y--;
-    if (this.input.down) this.velocity.y++;
-    if (this.input.left) this.velocity.x--;
-    if (this.input.right) this.velocity.x++;
-
-    const PLAYER_WIDTH = 32;
-    const PLAYER_HEIGHT = 32;
-    const HALF_PLAYER_WIDTH = PLAYER_WIDTH / 2;
-    const HALF_PLAYER_HEIGHT = PLAYER_HEIGHT / 2;
-    const centerX = Math.floor((this.position.x + HALF_PLAYER_WIDTH) / PLAYER_WIDTH);
-    const centerY = Math.floor((this.position.y + HALF_PLAYER_HEIGHT) / PLAYER_HEIGHT);
-
-    const block = game.world.blockAt(centerX, centerY, TileLayer.Action);
-
-    // TODO: this better
-    if (block === game.tileJson.id("arrow-up")) this.velocity.y = -2;
-    if (block === game.tileJson.id("arrow-down")) this.velocity.y = 2;
-    if (block === game.tileJson.id("arrow-left")) this.velocity.x = -2;
-    if (block === game.tileJson.id("arrow-right")) this.velocity.x = 2;
-
-    this.position.x += this.velocity.x * HALF_PLAYER_WIDTH;
-    this.position.y += this.velocity.y * HALF_PLAYER_HEIGHT;
-
-    if (this.position.x < 0) this.position.x = 0;
-    if (this.position.x > (game.world.size.width - 1) * PLAYER_WIDTH)
-      this.position.x = (game.world.size.width - 1) * PLAYER_WIDTH;
-
-    if (this.position.y < 0) this.position.y = 0;
-    if (this.position.y > (game.world.size.height - 1) * PLAYER_HEIGHT)
-      this.position.y = (game.world.size.height - 1) * PLAYER_HEIGHT;
-  }
-
-  destroy() {
-    // TODO: run code that needs to be run on deletion
-  }
-}
-
-export interface ReadOnlyWorld {
-  layerOfTopmostBlock(x: number, y: number): TileLayer;
-  blockAt(x: number, y: number, layer: TileLayer): number;
-}
-
-export class World implements ReadOnlyWorld {
-  protected state: number[][][];
-
-  private static emptyWorld(size: Size): number[][][] {
-    const state = [];
-
-    for (let idxLayer = TileLayer.Foreground; idxLayer <= TileLayer.Decoration; idxLayer++) {
-      const layer = [];
-      for (let idxY = 0; idxY < size.height; idxY++) {
-        const y = [];
-        for (let idxX = 0; idxX < size.width; idxX++) {
-          y[idxX] = 0;
-        }
-        layer[idxY] = y;
-      }
-      state[idxLayer] = layer;
-    }
-
-    return state;
-  }
-
-  private static placeBorder(state: number[][][], tileJson: TileRegistration, size: Size) {
-    const layer = state[TileLayer.Foreground];
-    const block = tileJson.id("basic-white");
-
-    const top = layer[0];
-    const bottom = layer[size.height - 1];
-
-    for (let x = 0; x < size.width; x++) {
-      top[x] = bottom[x] = block;
-    }
-
-    const rightEndOfWorld = size.width - 1;
-    for (let idxY = 1; idxY < size.height - 2; idxY++) {
-      const y = layer[idxY];
-      y[0] = y[rightEndOfWorld] = block;
-    }
-  }
-
-  constructor(protected readonly tileJson: TileRegistration, readonly size: Size) {
-    this.state = World.emptyWorld(size);
-  }
-
-  layerOfTopmostBlock(x: number, y: number): TileLayer {
-    // TODO: support decoration layers
-    // this is set to background because server doesn't send decoration layer
-    for (let layer = TileLayer.Background; layer >= TileLayer.Foreground; layer--) {
-      if (this.state[layer][y][x] !== 0) return layer;
-    }
-    return TileLayer.Foreground;
-  }
-
-  blockAt(x: number, y: number, layer: TileLayer): number {
-    return this.state[layer][y][x];
-  }
-
-  placeBlock(author: Player, x: number, y: number, id: number, layer?: TileLayer) {
-    if (x < 0 || y < 0 || x >= this.size.width || y >= this.size.height) return;
-    layer = layer ?? this.tileJson.for(id).layer;
-    this.state[layer][y][x] = id;
-    this.onPlace(layer, y, x, id);
-  }
-
-  placeLine(author: Player, x1: number, y1: number, x2: number, y2: number, id: number, layerParam?: number) {
-    const layer = layerParam ?? this.tileJson.for(id).layer;
-    const stateLayer = this.state[layer];
-    bresenhamsLine(x1, y1, x2, y2, (x, y) => {
-      if (x < 0 || y < 0 || x >= this.size.width || y >= this.size.height) return;
-      stateLayer[y][x] = id;
-      this.onPlace(layer, y, x, id);
-    });
-  }
-
-  load(blocks: number[][][]) {
-    this.state = blocks;
-  }
-
-  clear() {
-    this.state = World.emptyWorld(this.size);
-    World.placeBorder(this.state, this.tileJson, this.size);
-  }
-
-  onSave(author: Player) {}
-
-  onLoad(author: Player, blocks: number[][][]) {
-    this.load(blocks);
-  }
-
-  onClear(author: Player) {
-    this.clear();
-  }
-
-  onPlace(layer: TileLayer, y: number, x: number, id: number) {}
-}
-
-interface Message {
-  // id is for react
-  id: number;
-  time: Date;
-  // TODO: don't use `Player` so that the GC can pick up old players, use a
-  // smaller version of `Player` that carries enough information about the
-  // player to render them (e.g. role/name is all)
-  sender: Player;
-  content: string;
-}
-
-export class Bullets {
-  private readonly bullets: Bullet[] = [];
-  private readonly B: BulletCtor;
-
-  constructor(constructor?: BulletCtor) {
-    this.B = constructor || Bullet;
-  }
-
-  spawn(at: Player, angle: number) {
-    // TODO: put bullet in front of gun
-    const bullet = new this.B(at.position.x, at.position.y, angle);
-    this.bullets.push(bullet);
-  }
-
-  [Symbol.iterator]() {
-    return this.bullets[Symbol.iterator]();
-  }
-}
-
-export class Chat {
-  private _atTimeCanSend: Date = new Date();
-  private readonly _messages: Message[] = [];
-  private _topId: number = 0;
-
-  get messages(): Iterable<Message> {
-    return this._messages;
-  }
-
-  get topId(): number {
-    return this._topId;
-  }
-
-  add(time: Date, sender: Player, message: string) {
-    this._messages.push({ id: this._topId++, time, sender, content: message });
-  }
-
-  ratelimitFor(durationMs: number) {
-    this._atTimeCanSend = new Date();
-    this._atTimeCanSend.setTime(this._atTimeCanSend.getTime() + durationMs);
-  }
-}
-
-export interface Network {
-  update(game: Game): void;
-}
-
-export class Players {
-  private readonly _map: Map<number, Player> = new Map();
-  private readonly P: PlayerCtor;
-
-  constructor(constructor?: PlayerCtor) {
-    this.P = constructor || Player;
-  }
-
-  getPlayer(id: number): Player {
-    const player = this._map.get(id);
-    if (!player) throw new Error(`getPlayer failed with id ${id}`);
-    return player;
-  }
-
-  addPlayer(joinInfo: ZSPlayerJoin): Player {
-    const player = new this.P(joinInfo.playerId, joinInfo.username, joinInfo.isGuest);
-
-    player.role = joinInfo.role;
-    player.position = joinInfo.joinLocation;
-
-    if (joinInfo.hasGun) {
-      player.pickupGun();
-
-      if (joinInfo.gunEquipped) {
-        player.holdGun(true);
-      }
-    }
-
-    this._map.set(joinInfo.playerId, player);
-    return player;
-  }
-
-  removePlayer(id: number): Player {
-    const player = this.getPlayer(id);
-    player.destroy();
-    this._map.delete(id);
-    return player;
-  }
-
-  [Symbol.iterator]() {
-    return this._map.values();
-  }
-}
+import type { ZSPacket, ZSInit, ZSEvent, ZWorldAction } from "@smiley-face-game/api/packets";
+import { ZWorldActionKindReply } from "@smiley-face-game/api/types";
+import Player from "./Player";
+import World from "./World";
+import Players from "./Players";
+import Chat from "./Chat";
+import Bullets from "./Bullets";
 
 type GameFactory = () => [Bullets, Chat, Players, World];
-
-export interface Display {
-  draw(game: Game): void;
-}
 
 /**
  * `Game` class contains everything relevant for data, making it perfectly suitable
  * to use as a headless instance of a game (and thus, testing).
  */
-export class Game {
+export default class Game {
   readonly bullets: Bullets;
   readonly chat: Chat;
   readonly players: Players;
