+++ packages/client/src/game/GunState.ts
@@ -1,0 +1,6 @@
+enum GunState {
+  None,
+  Carrying,
+  Held,
+}
+export default GunState;
