+++ packages/client/src/game/Inputs.ts
@@ -1,0 +1,7 @@
+export default interface Inputs {
+  jump: boolean;
+  up: boolean;
+  down: boolean;
+  left: boolean;
+  right: boolean;
+}
