+++ packages/client/src/game/Keyboard.ts
@@ -1,0 +1,50 @@
+import Player from "./Player";
+
+export default class Keyboard {
+  constructor(player: Player) {
+    // TODO: do this nicely? (inside of ClientGame)
+    document.addEventListener("keydown", (event) => {
+      const key = event.key.toLowerCase();
+
+      switch (key) {
+        case "w":
+          player.input.up = true;
+          break;
+        case "d":
+          player.input.right = true;
+          break;
+        case "a":
+          player.input.left = true;
+          break;
+        case "s":
+          player.input.down = true;
+          break;
+        case "space":
+          player.input.jump = true;
+          break;
+      }
+    });
+
+    document.addEventListener("keyup", (event) => {
+      const key = event.key.toLowerCase();
+
+      switch (key) {
+        case "w":
+          player.input.up = !true;
+          break;
+        case "d":
+          player.input.right = !true;
+          break;
+        case "a":
+          player.input.left = !true;
+          break;
+        case "s":
+          player.input.down = !true;
+          break;
+        case "space":
+          player.input.jump = !true;
+          break;
+      }
+    });
+  }
+}
