+++ packages/client/src/game/Message.ts
@@ -1,0 +1,12 @@
+import Player from "./Player";
+
+export default interface Message {
+  // id is for react
+  id: number;
+  time: Date;
+  // TODO: don't use `Player` so that the GC can pick up old players, use a
+  // smaller version of `Player` that carries enough information about the
+  // player to render them (e.g. role/name is all)
+  sender: Player;
+  content: string;
+}
