+++ packages/client/src/game/MouseState.ts
@@ -1,0 +1,8 @@
+enum MouseState {
+  None,
+  Place,
+  Erase,
+  WasPlacingNowErasing,
+  WasErasingNowPlacing,
+}
+export default MouseState;
