+++ packages/client/src/game/Network.ts
@@ -1,0 +1,5 @@
+import Game from "./Game";
+
+export default interface Network {
+  update(game: Game): void;
+}
