+++ packages/client/src/game/PhysicsObject.ts
@@ -1,0 +1,9 @@
+import Position from "./Position";
+import Velocity from "./Velocity";
+import Game from "./Game";
+
+export default interface PhysicsObject {
+  position: Position;
+  velocity: Velocity;
+  tick?: (game: Game) => void;
+}
