+++ packages/client/src/game/PlayerCtor.ts
@@ -1,0 +1,5 @@
+import Player from "./Player";
+
+export default interface PlayerCtor {
+  new (id: number, username: string, isGuest: boolean): Player;
+}
