+++ packages/client/src/game/Position.ts
@@ -1,0 +1,4 @@
+export default interface Position {
+  x: number;
+  y: number;
+}
