+++ packages/client/src/game/ReadOnlyWorld.ts
@@ -1,0 +1,6 @@
+import { TileLayer } from "@smiley-face-game/api/types";
+
+export default interface ReadOnlyWorld {
+  layerOfTopmostBlock(x: number, y: number): TileLayer;
+  blockAt(x: number, y: number, layer: TileLayer): number;
+}
