+++ packages/client/src/game/Size.ts
@@ -1,0 +1,4 @@
+export default interface Size {
+  width: number;
+  height: number;
+}
