+++ packages/client/src/game/Velocity.ts
@@ -1,0 +1,4 @@
+import Position from "./Position";
+
+type Velocity = Position;
+export default Velocity;
