+++ packages/client/src/game/defaultInputs.ts
@@ -1,0 +1,5 @@
+import Inputs from "./Inputs";
+
+export default function defaultInputs(): Inputs {
+  return { jump: false, up: false, down: false, left: false, right: false };
+}
